Reader mode pages back too — which is what was asked for
The windowing went into the Trace pane, and the Trace pane is not where anyone
reads a conversation. Reader mode is, and it still fetched the last 400 turns in
one request and stopped there: 702 KB on a 19 MB session, "1,020 earlier
messages are not shown", and no way to reach them. It also re-fetched all 400
every three seconds while the agent worked, each one a full re-parse of the
transcript on the server.

Reader mode now opens on a window of the end, pages backwards when you scroll to
the top, and follows the end with `after=`. Measured on the 19 MB / 1,420-turn
session: the open drops 702 KB → 84 KB, and the poll while an agent works drops
702 KB + a whole-file parse → 359 bytes and a bounded read. Paging back reaches
the first turn in 34 windows.

The paging itself moved to lib/traceWindows.ts and both surfaces use it. Writing
it twice was not an option — the cursor handling, the single-flight guard, the
generation stamp that stops a load outliving its source, the gap and blocked
cases, are all things a review round found rather than things that were obvious.
What stays with each surface is presentation: the Trace pane measures rows and
virtualizes; reader mode renders exchanges and does not.

Anchoring had to differ, and that is the interesting part. The Trace pane pins a
row by its own geometry. Reader mode cannot: its unit is an exchange, and an
exchange at the top of the list is a fragment whose prompt was in the window we
had not read yet — when that window arrives the two become ONE exchange, so the
element the reader was looking at ceases to exist. What does not change is that
everything new is added above, so it restores the distance to the bottom
instead. Verified by what is under the middle of the pane rather than by element
identity: same text before and after.

No virtualization in reader mode yet, because measurement says it is not needed:
the whole 1,420-turn conversation paged in is 2,719 DOM nodes scrolling at a
16.6 ms median frame (worst 19.3 ms). Windowing by exchange (spec §10.7) is
still the right next step for a session an order of magnitude longer.

Two bugs this shook out, both mine, both caught by the browser checks rather than
by reasoning:

- the shared hook has to say whether turns arrived at the FRONT or the BACK. A
  single "the array grew" path prepended height slots for an append too, which
  silently misaligned the Trace pane's height model on every live poll.
- a prepend's scroll anchor must outrank a measurement's. The ResizeObserver
  captures an anchor to keep the view still across a re-measure; landing between
  a prepend's capture and its restore, it overwrote the anchor with the OLD
  indexing, and the view was restored a whole window away from where the reader
  was — which then triggered another load, and another.
