+++ web/src/conversation.css
@@ -166,6 +166,10 @@ mark.cx-hit.on { background: var(--accent); color: var(--panel); }
 .cxv-body { flex: 1; min-height: 0; overflow-y: auto; background: var(--panel); padding: 4px 14px 30px; }
 .cxv-col { display: flex; flex-direction: column; min-width: 0; }
 .cxv-msg { font-size: 11px; color: var(--muted); padding: 6px 0; }
+/* The line above the oldest loaded exchange. Fixed height on purpose: it sits
+   above everything in the column, so a line that grew or vanished when the
+   reader reached the start would shove the whole conversation. */
+.cxv-top { height: 1.5em; line-height: 1.5em; padding: 0; text-align: center; opacity: 0.75; }
 .cxv-foot {
   display: flex; align-items: center; gap: 8px; flex: none;
   padding: 5px 10px; border-top: 1px solid var(--border); background: var(--panel);
